refactor(core): move key encoding into engine
