+++ crates/core/src/lib.rs
@@ -7,6 +7,7 @@ mod events;
 pub use events::{CoreEvent, DisplayClearMode, IngestDegradeReason, LineClearMode};
 
 mod grid;
+mod input;
 mod parser;
 pub use parser::ShellMarkerKind;
 mod render_mode;
@@ -17,6 +18,9 @@ pub use cursor::Cursor;
 pub use grid::{
     ANSI_PALETTE, Attrs, CELL_HEIGHT, CELL_WIDTH, Cell, Color, DEFAULT_BG, DEFAULT_FG, Grid,
     UnderlineStyle, color_to_u32,
+};
+pub use input::{
+    RuntimeKey, RuntimeKeyEvent, RuntimeKeyModifiers, TerminalModeFlags, encode_runtime_key_event,
 };
 // Keep parser module private while preserving a narrow tooling surface for direct fuzz harnesses.
 pub use parser::Parser;
