+++ crates/features/interaction/src/input.rs
@@ -2,73 +2,32 @@
 // Copyright (C) 2026 Danil Silantyev, Global CEO NDDev. on.nddev.it.com (OpenNetwork)
 
 use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
+use rldyourterm_core::{
+    RuntimeKey, RuntimeKeyEvent, RuntimeKeyModifiers, TerminalModeFlags, encode_runtime_key_event,
+};
 use winit::event::KeyEvent as WinitKeyEvent;
 use winit::keyboard::{Key, ModifiersState, NamedKey};
 
-use crate::key_encoding::{
-    cursor_key, encode_ctrl_letter, fkey_ss3_modified, tilde_modified, xterm_modifier_param,
-};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum RuntimeKey {
-    Character(char),
-    Enter,
-    Tab,
-    Backspace,
-    Escape,
-    Up,
-    Down,
-    Right,
-    Left,
-    Home,
-    End,
-    Delete,
-    Insert,
-    PageUp,
-    PageDown,
-    F(u8),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct RuntimeKeyModifiers {
-    pub shift: bool,
-    pub alt: bool,
-    pub control: bool,
-    pub super_key: bool,
-}
-
-impl RuntimeKeyModifiers {
-    fn from_crossterm(modifiers: KeyModifiers) -> Self {
-        Self {
-            shift: modifiers.contains(KeyModifiers::SHIFT),
-            alt: modifiers.contains(KeyModifiers::ALT),
-            control: modifiers.contains(KeyModifiers::CONTROL),
-            super_key: modifiers.contains(KeyModifiers::SUPER),
-        }
-    }
-
-    fn from_winit(modifiers: ModifiersState) -> Self {
-        Self {
-            shift: modifiers.shift_key(),
-            alt: modifiers.alt_key(),
-            control: modifiers.control_key(),
-            super_key: modifiers.super_key(),
-        }
-    }
-
-    fn xterm_modifier_param(self) -> u8 {
-        xterm_modifier_param(self.shift, self.alt, self.control)
+fn runtime_key_modifiers_from_crossterm(modifiers: KeyModifiers) -> RuntimeKeyModifiers {
+    RuntimeKeyModifiers {
+        shift: modifiers.contains(KeyModifiers::SHIFT),
+        alt: modifiers.contains(KeyModifiers::ALT),
+        control: modifiers.contains(KeyModifiers::CONTROL),
+        super_key: modifiers.contains(KeyModifiers::SUPER),
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct RuntimeKeyEvent {
-    pub key: RuntimeKey,
-    pub modifiers: RuntimeKeyModifiers,
+fn runtime_key_modifiers_from_winit(modifiers: ModifiersState) -> RuntimeKeyModifiers {
+    RuntimeKeyModifiers {
+        shift: modifiers.shift_key(),
+        alt: modifiers.alt_key(),
+        control: modifiers.control_key(),
+        super_key: modifiers.super_key(),
+    }
 }
 
 pub fn runtime_key_event_from_crossterm(key_event: KeyEvent) -> Option<RuntimeKeyEvent> {
-    let mut modifiers = RuntimeKeyModifiers::from_crossterm(key_event.modifiers);
+    let mut modifiers = runtime_key_modifiers_from_crossterm(key_event.modifiers);
     let key = match key_event.code {
         KeyCode::Char(ch) => RuntimeKey::Character(ch),
         KeyCode::Enter => RuntimeKey::Enter,
@@ -105,7 +64,7 @@ pub fn runtime_key_event_from_winit(
 ) -> Option<RuntimeKeyEvent> {
     Some(RuntimeKeyEvent {
         key: runtime_key_from_winit(key)?,
-        modifiers: RuntimeKeyModifiers::from_winit(modifiers),
+        modifiers: runtime_key_modifiers_from_winit(modifiers),
     })
 }
 
@@ -135,7 +94,7 @@ pub fn is_runtime_palette_shortcut_winit(key: Key<&str>, modifiers: ModifiersSta
     runtime_key_from_winit_borrowed(key)
         .map(|key| RuntimeKeyEvent {
             key,
-            modifiers: RuntimeKeyModifiers::from_winit(modifiers),
+            modifiers: runtime_key_modifiers_from_winit(modifiers),
         })
         .is_some_and(is_runtime_palette_shortcut)
 }
@@ -156,158 +115,6 @@ pub fn is_local_shutdown_key_winit(event: &WinitKeyEvent, modifiers: ModifiersSt
     runtime_key_event_from_winit(&event.logical_key, modifiers).is_some_and(is_local_shutdown_key)
 }
 
-#[derive(Debug, Clone, Copy)]
-pub struct TerminalModeFlags {
-    pub application_cursor_keys: bool,
-    pub kitty_keyboard_flags: u16,
-    pub meta_sends_escape: bool,
-    pub alt_sends_escape: bool,
-}
-
-impl Default for TerminalModeFlags {
-    fn default() -> Self {
-        Self {
-            application_cursor_keys: false,
-            kitty_keyboard_flags: 0,
-            meta_sends_escape: true,
-            alt_sends_escape: false,
-        }
-    }
-}
-
-impl TerminalModeFlags {
-    fn alt_modifier_sends_escape(self) -> bool {
-        // The runtime input model exposes Alt but not a distinct Meta modifier bit.
-        // Treat Alt as the shared Meta/Alt-like modifier and let the terminal modes
-        // decide whether that modifier should prefix ESC.
-        self.meta_sends_escape || self.alt_sends_escape
-    }
-}
-
-pub fn encode_runtime_key_event(
-    key_event: RuntimeKeyEvent,
-    modes: TerminalModeFlags,
-) -> Option<Vec<u8>> {
-    if modes.kitty_keyboard_flags > 0 {
-        return encode_kitty_key_event(key_event, modes);
-    }
-    encode_legacy_key_event(key_event, modes)
-}
-
-fn encode_legacy_key_event(
-    key_event: RuntimeKeyEvent,
-    modes: TerminalModeFlags,
-) -> Option<Vec<u8>> {
-    let modifiers = key_event.modifiers;
-    let mod_param = modifiers.xterm_modifier_param();
-    let has_mod = mod_param > 1;
-    let app_cursor = modes.application_cursor_keys;
-
-    match key_event.key {
-        RuntimeKey::Enter => Some(vec![b'\r']),
-        RuntimeKey::Backspace
-            if modifiers.alt && !modifiers.control && modes.alt_modifier_sends_escape() =>
-        {
-            Some(b"\x1b\x7f".to_vec())
-        }
-        RuntimeKey::Backspace => Some(vec![0x7f]),
-        RuntimeKey::Tab if modifiers.shift && !modifiers.alt && !modifiers.control => {
-            Some(b"\x1b[Z".to_vec())
-        }
-        RuntimeKey::Tab if !modifiers.alt && !modifiers.control => Some(vec![b'\t']),
-        RuntimeKey::Escape => Some(vec![0x1b]),
-        RuntimeKey::Up => Some(cursor_key(b'A', mod_param, has_mod, app_cursor)),
-        RuntimeKey::Down => Some(cursor_key(b'B', mod_param, has_mod, app_cursor)),
-        RuntimeKey::Right => Some(cursor_key(b'C', mod_param, has_mod, app_cursor)),
-        RuntimeKey::Left => Some(cursor_key(b'D', mod_param, has_mod, app_cursor)),
-        RuntimeKey::Home => Some(cursor_key(b'H', mod_param, has_mod, app_cursor)),
-        RuntimeKey::End => Some(cursor_key(b'F', mod_param, has_mod, app_cursor)),
-        RuntimeKey::Delete => Some(tilde_modified(3, mod_param, has_mod)),
-        RuntimeKey::Insert => Some(tilde_modified(2, mod_param, has_mod)),
-        RuntimeKey::PageUp => Some(tilde_modified(5, mod_param, has_mod)),
-        RuntimeKey::PageDown => Some(tilde_modified(6, mod_param, has_mod)),
-        RuntimeKey::F(1) => Some(fkey_ss3_modified(b'P', mod_param, has_mod)),
-        RuntimeKey::F(2) => Some(fkey_ss3_modified(b'Q', mod_param, has_mod)),
-        RuntimeKey::F(3) => Some(fkey_ss3_modified(b'R', mod_param, has_mod)),
-        RuntimeKey::F(4) => Some(fkey_ss3_modified(b'S', mod_param, has_mod)),
-        RuntimeKey::F(5) => Some(tilde_modified(15, mod_param, has_mod)),
-        RuntimeKey::F(6) => Some(tilde_modified(17, mod_param, has_mod)),
-        RuntimeKey::F(7) => Some(tilde_modified(18, mod_param, has_mod)),
-        RuntimeKey::F(8) => Some(tilde_modified(19, mod_param, has_mod)),
-        RuntimeKey::F(9) => Some(tilde_modified(20, mod_param, has_mod)),
-        RuntimeKey::F(10) => Some(tilde_modified(21, mod_param, has_mod)),
-        RuntimeKey::F(11) => Some(tilde_modified(23, mod_param, has_mod)),
-        RuntimeKey::F(12) => Some(tilde_modified(24, mod_param, has_mod)),
-        RuntimeKey::F(13) => Some(tilde_modified(25, mod_param, has_mod)),
-        RuntimeKey::F(14) => Some(tilde_modified(26, mod_param, has_mod)),
-        RuntimeKey::F(15) => Some(tilde_modified(28, mod_param, has_mod)),
-        RuntimeKey::F(16) => Some(tilde_modified(29, mod_param, has_mod)),
-        RuntimeKey::F(17) => Some(tilde_modified(31, mod_param, has_mod)),
-        RuntimeKey::F(18) => Some(tilde_modified(32, mod_param, has_mod)),
-        RuntimeKey::F(19) => Some(tilde_modified(33, mod_param, has_mod)),
-        RuntimeKey::F(20) => Some(tilde_modified(34, mod_param, has_mod)),
-        RuntimeKey::Character(ch) if modifiers.control => {
-            encode_ctrl_letter(ch).map(|code| vec![code])
-        }
-        RuntimeKey::Character(ch)
-            if modifiers.alt
-                && !modifiers.control
-                && !modifiers.super_key
-                && modes.alt_modifier_sends_escape() =>
-        {
-            let mut bytes = vec![0x1b];
-            bytes.extend_from_slice(ch.to_string().as_bytes());
-            Some(bytes)
-        }
-        RuntimeKey::Character(ch) if !modifiers.control && !modifiers.super_key => {
-            Some(ch.to_string().into_bytes())
-        }
-        RuntimeKey::Character(_) => None,
-        RuntimeKey::Tab => None,
-        RuntimeKey::F(_) => None,
-    }
-}
-
-fn encode_kitty_key_event(key_event: RuntimeKeyEvent, modes: TerminalModeFlags) -> Option<Vec<u8>> {
-    let modifiers = key_event.modifiers;
-    let kitty_mod = kitty_modifier_param(modifiers);
-    let has_mod = kitty_mod > 1;
-
-    match key_event.key {
-        RuntimeKey::Enter => Some(kitty_csi_u(13, kitty_mod, has_mod)),
-        RuntimeKey::Tab => Some(kitty_csi_u(9, kitty_mod, has_mod)),
-        RuntimeKey::Backspace => Some(kitty_csi_u(127, kitty_mod, has_mod)),
-        RuntimeKey::Escape => Some(kitty_csi_u(27, kitty_mod, has_mod)),
-        RuntimeKey::Character(ch) => Some(kitty_csi_u(u32::from(ch), kitty_mod, has_mod)),
-        RuntimeKey::Up
-        | RuntimeKey::Down
-        | RuntimeKey::Right
-        | RuntimeKey::Left
-        | RuntimeKey::Home
-        | RuntimeKey::End
-        | RuntimeKey::Delete
-        | RuntimeKey::Insert
-        | RuntimeKey::PageUp
-        | RuntimeKey::PageDown
-        | RuntimeKey::F(_) => encode_legacy_key_event(key_event, modes),
-    }
-}
-
-fn kitty_modifier_param(modifiers: RuntimeKeyModifiers) -> u8 {
-    1 + u8::from(modifiers.shift)
-        + (u8::from(modifiers.alt) << 1)
-        + (u8::from(modifiers.control) << 2)
-        + (u8::from(modifiers.super_key) << 3)
-}
-
-fn kitty_csi_u(codepoint: u32, mod_param: u8, has_mod: bool) -> Vec<u8> {
-    if has_mod {
-        format!("\x1b[{codepoint};{mod_param}u").into_bytes()
-    } else {
-        format!("\x1b[{codepoint}u").into_bytes()
-    }
-}
-
 pub fn encode_crossterm_key_event(
     key_event: KeyEvent,
     modes: TerminalModeFlags,
@@ -321,7 +128,7 @@ pub fn encode_winit_key_event(
     modifiers: ModifiersState,
     modes: TerminalModeFlags,
 ) -> Option<Vec<u8>> {
-    let modifiers = RuntimeKeyModifiers::from_winit(modifiers);
+    let modifiers = runtime_key_modifiers_from_winit(modifiers);
     if modes.kitty_keyboard_flags == 0
         && let Some(bytes) = encode_winit_text_bytes(event.text.as_deref(), modifiers, modes)
     {
@@ -418,12 +225,11 @@ fn runtime_key_from_winit_ref(key: Key<&str>) -> Option<RuntimeKey> {
 #[cfg(test)]
 mod tests {
     use super::{
-        RuntimeKey, RuntimeKeyEvent, RuntimeKeyModifiers, TerminalModeFlags,
-        encode_crossterm_key_event, encode_runtime_key_event, encode_winit_text_bytes,
-        is_runtime_palette_shortcut_winit, runtime_key_from_crossterm,
-        runtime_key_from_winit_borrowed,
+        encode_crossterm_key_event, encode_winit_text_bytes, is_runtime_palette_shortcut_winit,
+        runtime_key_from_crossterm, runtime_key_from_winit_borrowed,
     };
     use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
+    use rldyourterm_core::{RuntimeKey, RuntimeKeyModifiers, TerminalModeFlags};
     use winit::keyboard::{Key, ModifiersState, NamedKey, SmolStr};
 
     fn encode_key(key_event: KeyEvent) -> Option<Vec<u8>> {
@@ -538,320 +344,6 @@ mod tests {
         assert_eq!(
             encode_key(KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE)),
             Some(vec![b'\r'])
-        );
-    }
-
-    #[test]
-    fn application_cursor_keys_sends_ss3_for_arrows() {
-        let modes = TerminalModeFlags {
-            application_cursor_keys: true,
-            ..Default::default()
-        };
-        let no_mods = RuntimeKeyModifiers {
-            shift: false,
-            alt: false,
-            control: false,
-            super_key: false,
-        };
-        let ev = RuntimeKeyEvent {
-            key: RuntimeKey::Up,
-            modifiers: no_mods,
-        };
-        assert_eq!(
-            encode_runtime_key_event(ev, modes),
-            Some(b"\x1bOA".to_vec())
-        );
-
-        let ev = RuntimeKeyEvent {
-            key: RuntimeKey::Down,
-            modifiers: no_mods,
-        };
-        assert_eq!(
-            encode_runtime_key_event(ev, modes),
-            Some(b"\x1bOB".to_vec())
-        );
-
-        let ev = RuntimeKeyEvent {
-            key: RuntimeKey::Right,
-            modifiers: no_mods,
-        };
-        assert_eq!(
-            encode_runtime_key_event(ev, modes),
-            Some(b"\x1bOC".to_vec())
-        );
-
-        let ev = RuntimeKeyEvent {
-            key: RuntimeKey::Left,
-            modifiers: no_mods,
-        };
-        assert_eq!(
-            encode_runtime_key_event(ev, modes),
-            Some(b"\x1bOD".to_vec())
-        );
-    }
-
-    #[test]
-    fn application_cursor_keys_uses_csi_when_modifier_present() {
-        let modes = TerminalModeFlags {
-            application_cursor_keys: true,
-            ..Default::default()
-        };
-        let ctrl = RuntimeKeyModifiers {
-            shift: false,
-            alt: false,
-            control: true,
-            super_key: false,
-        };
-        let ev = RuntimeKeyEvent {
-            key: RuntimeKey::Up,
-            modifiers: ctrl,
-        };
-        assert_eq!(
-            encode_runtime_key_event(ev, modes),
-            Some(b"\x1b[1;5A".to_vec())
-        );
-    }
-
-    #[test]
-    fn kitty_mode_encodes_regular_keys_as_csi_u() {
-        let modes = TerminalModeFlags {
-            application_cursor_keys: false,
-            kitty_keyboard_flags: 1,
-            ..Default::default()
-        };
-        let key_event = RuntimeKeyEvent {
-            key: RuntimeKey::Character('a'),
-            modifiers: RuntimeKeyModifiers {
-                shift: false,
-                alt: false,
-                control: false,
-                super_key: false,
-            },
-        };
-        assert_eq!(
-            encode_runtime_key_event(key_event, modes),
-            Some(b"\x1b[97u".to_vec())
-        );
-    }
-
-    #[test]
-    fn kitty_mode_encodes_enter_as_csi_13_u() {
-        let modes = TerminalModeFlags {
-            application_cursor_keys: false,
-            kitty_keyboard_flags: 1,
-            ..Default::default()
-        };
-        let key_event = RuntimeKeyEvent {
-            key: RuntimeKey::Enter,
-            modifiers: RuntimeKeyModifiers {
-                shift: false,
-                alt: false,
-                control: false,
-                super_key: false,
-            },
-        };
-        assert_eq!(
-            encode_runtime_key_event(key_event, modes),
-            Some(b"\x1b[13u".to_vec())
-        );
-    }
-
-    #[test]
-    fn kitty_mode_encodes_tab_as_csi_9_u() {
-        let modes = TerminalModeFlags {
-            application_cursor_keys: false,
-            kitty_keyboard_flags: 1,
-            ..Default::default()
-        };
-        let key_event = RuntimeKeyEvent {
-            key: RuntimeKey::Tab,
-            modifiers: RuntimeKeyModifiers {
-                shift: false,
-                alt: false,
-                control: false,
-                super_key: false,
-            },
-        };
-        assert_eq!(
-            encode_runtime_key_event(key_event, modes),
-            Some(b"\x1b[9u".to_vec())
-        );
-    }
-
-    #[test]
-    fn kitty_mode_encodes_backspace_as_csi_127_u() {
-        let modes = TerminalModeFlags {
-            application_cursor_keys: false,
-            kitty_keyboard_flags: 1,
-            ..Default::default()
-        };
-        let key_event = RuntimeKeyEvent {
-            key: RuntimeKey::Backspace,
-            modifiers: RuntimeKeyModifiers {
-                shift: false,
-                alt: false,
-                control: false,
-                super_key: false,
-            },
-        };
-        assert_eq!(
-            encode_runtime_key_event(key_event, modes),
-            Some(b"\x1b[127u".to_vec())
-        );
-    }
-
-    #[test]
-    fn kitty_mode_encodes_escape_as_csi_27_u() {
-        let modes = TerminalModeFlags {
-            application_cursor_keys: false,
-            kitty_keyboard_flags: 1,
-            ..Default::default()
-        };
-        let key_event = RuntimeKeyEvent {
-            key: RuntimeKey::Escape,
-            modifiers: RuntimeKeyModifiers {
-                shift: false,
-                alt: false,
-                control: false,
-                super_key: false,
-            },
-        };
-        assert_eq!(
-            encode_runtime_key_event(key_event, modes),
-            Some(b"\x1b[27u".to_vec())
-        );
-    }
-
-    #[test]
-    fn kitty_mode_encodes_ctrl_a_with_modifier() {
-        let modes = TerminalModeFlags {
-            application_cursor_keys: false,
-            kitty_keyboard_flags: 1,
-            ..Default::default()
-        };
-        let key_event = RuntimeKeyEvent {
-            key: RuntimeKey::Character('a'),
-            modifiers: RuntimeKeyModifiers {
-                shift: false,
-                alt: false,
-                control: true,
-                super_key: false,
-            },
-        };
-        assert_eq!(
-            encode_runtime_key_event(key_event, modes),
-            Some(b"\x1b[97;5u".to_vec())
-        );
-    }
-
-    #[test]
-    fn kitty_mode_encodes_shift_tab_with_modifier() {
-        let modes = TerminalModeFlags {
-            application_cursor_keys: false,
-            kitty_keyboard_flags: 1,
-            ..Default::default()
-        };
-        let key_event = RuntimeKeyEvent {
-            key: RuntimeKey::Tab,
-            modifiers: RuntimeKeyModifiers {
-                shift: true,
-                alt: false,
-                control: false,
-                super_key: false,
-            },
-        };
-        assert_eq!(
-            encode_runtime_key_event(key_event, modes),
-            Some(b"\x1b[9;2u".to_vec())
-        );
-    }
-
-    #[test]
-    fn kitty_mode_arrows_use_legacy_encoding() {
-        let modes = TerminalModeFlags {
-            application_cursor_keys: false,
-            kitty_keyboard_flags: 1,
-            ..Default::default()
-        };
-        let key_event = RuntimeKeyEvent {
-            key: RuntimeKey::Up,
-            modifiers: RuntimeKeyModifiers {
-                shift: false,
-                alt: false,
-                control: false,
-                super_key: false,
-            },
-        };
-        assert_eq!(
-            encode_runtime_key_event(key_event, modes),
-            Some(b"\x1b[A".to_vec())
-        );
-    }
-
-    #[test]
-    fn kitty_mode_disabled_uses_legacy() {
-        let modes = TerminalModeFlags::default();
-        let key_event = RuntimeKeyEvent {
-            key: RuntimeKey::Character('a'),
-            modifiers: RuntimeKeyModifiers {
-                shift: false,
-                alt: false,
-                control: false,
-                super_key: false,
-            },
-        };
-        assert_eq!(encode_runtime_key_event(key_event, modes), Some(vec![b'a']));
-    }
-
-    #[test]
-    fn legacy_alt_backspace_respects_escape_modes() {
-        let disabled = TerminalModeFlags {
-            meta_sends_escape: false,
-            alt_sends_escape: false,
-            ..Default::default()
-        };
-        let enabled = TerminalModeFlags {
-            meta_sends_escape: false,
-            alt_sends_escape: true,
-            ..Default::default()
-        };
-        let key_event = RuntimeKeyEvent {
-            key: RuntimeKey::Backspace,
-            modifiers: RuntimeKeyModifiers {
-                alt: true,
-                ..Default::default()
-            },
-        };
-
-        assert_eq!(
-            encode_runtime_key_event(key_event, disabled),
-            Some(vec![0x7f])
-        );
-        assert_eq!(
-            encode_runtime_key_event(key_event, enabled),
-            Some(b"\x1b\x7f".to_vec())
-        );
-    }
-
-    #[test]
-    fn kitty_mode_super_modifier_encoded() {
-        let modes = TerminalModeFlags {
-            application_cursor_keys: false,
-            kitty_keyboard_flags: 1,
-            ..Default::default()
-        };
-        let key_event = RuntimeKeyEvent {
-            key: RuntimeKey::Character('a'),
-            modifiers: RuntimeKeyModifiers {
-                shift: false,
-                alt: false,
-                control: false,
-                super_key: true,
-            },
-        };
-        assert_eq!(
-            encode_runtime_key_event(key_event, modes),
-            Some(b"\x1b[97;9u".to_vec())
         );
     }
 }
