+++ crates/features/interaction/src/lib.rs
@@ -2,18 +2,16 @@
 // Copyright (C) 2026 Danil Silantyev, Global CEO NDDev. on.nddev.it.com (OpenNetwork)
 
 mod input;
-mod key_encoding;
 mod mouse;
 mod palette;
 mod state;
 
 pub use input::{
-    RuntimeKey, RuntimeKeyEvent, RuntimeKeyModifiers, TerminalModeFlags,
-    encode_crossterm_key_event, encode_runtime_key_event, encode_winit_key_event,
-    is_local_shutdown_key, is_local_shutdown_key_crossterm, is_local_shutdown_key_winit,
-    is_runtime_palette_shortcut, is_runtime_palette_shortcut_crossterm,
-    is_runtime_palette_shortcut_winit, runtime_key_event_from_crossterm,
-    runtime_key_event_from_winit, runtime_key_from_winit_borrowed,
+    encode_crossterm_key_event, encode_winit_key_event, is_local_shutdown_key,
+    is_local_shutdown_key_crossterm, is_local_shutdown_key_winit, is_runtime_palette_shortcut,
+    is_runtime_palette_shortcut_crossterm, is_runtime_palette_shortcut_winit,
+    runtime_key_event_from_crossterm, runtime_key_event_from_winit,
+    runtime_key_from_winit_borrowed,
 };
 pub use mouse::{
     PRIMARY_MOUSE_BUTTON_CODE, active_mouse_button_code, encode_mouse_event,
@@ -24,5 +22,8 @@ pub use palette::{
     RuntimePaletteDecision, RuntimePaletteDispatchResult, RuntimePaletteView,
     dispatch_runtime_palette_command, handle_runtime_palette_key_input, runtime_palette_info_line,
     runtime_palette_status_line, toggle_runtime_palette,
+};
+pub use rldyourterm_core::{
+    RuntimeKey, RuntimeKeyEvent, RuntimeKeyModifiers, TerminalModeFlags, encode_runtime_key_event,
 };
 pub use state::{GridPoint, InteractionState, PointerState, SelectionRange};
